+++ app/src/main/res/layout/activity_menu.xml
@@ -31,5 +31,18 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:navGraph="@navigation/mobile_navigation" />
+    <FrameLayout
+        android:id="@+id/fragment_container"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1"
+        />
+    <!-- res/menu/bottom_nav_menu.xml -->
+    <item
+        android:id="@+id/navigation_beEmpowered"
+        android:icon="@drawable/ic_empower"
+        android:title="Be Empowered" />
+
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
