+++ app/src/main/res/layout/fragment_empower.xml
@@ -1,9 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
+    android:layout_height="wrap_content"
+    android:orientation="vertical"
     tools:context=".EmpowerFragment">
 
     <!-- TODO: Update blank fragment layout -->
@@ -12,20 +14,23 @@
         android:id="@+id/Item_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginEnd="133dp"
         android:ems="10"
-        android:hint="Name"
-        android:inputType="text"
-        tools:layout_editor_absoluteX="68dp"
-        tools:layout_editor_absoluteY="16dp" />
+        android:hint="Bio"
+        android:inputType="text" />
 
-    <EditText
-        android:id="@+id/quantity"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:ems="10"
-        android:hint="quantity"
-        android:inputType="text"
-        tools:layout_editor_absoluteX="45dp"
-        tools:layout_editor_absoluteY="87dp" />
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical">
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+        <Button
+            android:id="@+id/Donatebtn"
+            android:layout_width="66dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:text="Donate" />
+
+    </LinearLayout>
+
+</LinearLayout>
